requirements.txt
IPython
flask
google-generativeai
python-dotenv